TTOL-46 refactor: 파일명 변경
_hooks와 _api 폴더 내부의 index.ts를 각각 useOutsideClick.ts, searchLocation.ts로 변경
